+++ source/databricks/tests/integration/test_dummy.py
@@ -1,26 +1,25 @@
+# Copyright 2020 Energinet DataHub A/S
+#
+# Licensed under the Apache License, Version 2.0 (the "License2");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import asyncio
 import pytest
-import os
 import threading
 import time
-from pyspark import SparkConf
 from pyspark.sql import SparkSession, DataFrame
 from pyspark.sql.functions import col, lit, to_timestamp, explode
 from pyspark.sql.types import StructType, StructField, StringType, ArrayType, \
     DecimalType, IntegerType, TimestampType, BooleanType, BinaryType, LongType
-
-@pytest.fixture()
-def spark():
-    # spark.hadoop.fs.* for Azurite storage
-    spark_conf = SparkConf(loadDefaults=True) \
-        .set("spark.sql.session.timeZone", "UTC")
-        #.set("spark.hadoop.fs.defaultFS", "wasb://container@azurite") \
-        #.set("spark.hadoop.fs.azure.storage.emulator.account.name", "azurite")
-    return SparkSession \
-        .builder \
-        .config(conf=spark_conf) \
-        .config("spark.sql.streaming.schemaInference", True) \
-        .getOrCreate()
 
 
 class Task(threading.Thread):
@@ -31,18 +30,6 @@ class Task(threading.Thread):
     def run(self):
         """Target function of the thread class"""
         self.function()
-
-
-@pytest.fixture(scope="session")
-def azurite():
-    import subprocess
-    azurite_process = subprocess.Popen(args=["azurite-blob", "-l", ".azurite-files"], stdout=subprocess.PIPE)
-
-    # import os
-    # os.system("azurite-blob -l .azurite-files")
-    yield
-    azurite_process.terminate()
-    print("############ Azurite has terminated")
 
 
 """
@@ -56,7 +43,7 @@ def spark_job(spark):
     raw_stream =(spark
             .readStream
             #.schema(schema)
-            .json("/workspaces/geh-timeseries/source/databricks/tests/test_data*.json"))
+            .json("/workspaces/geh-timeseries/source/databricks/tests/integration/test_data*.json"))
 
     query = (raw_stream
         .writeStream
@@ -71,6 +58,7 @@ def spark_job(spark):
 @pytest.fixture
 def my_job_function(spark: SparkSession, azurite):
     return lambda: spark_job(spark)
+
 
 """
 def test_spark(my_job_function):
@@ -92,7 +80,7 @@ async def job_task(spark):
     raw_stream =(spark
         .readStream
         #.schema(schema)
-        .json("/workspaces/geh-timeseries/source/databricks/tests/test_data*.json"))
+        .json("/workspaces/geh-timeseries/source/databricks/tests/integration/test_data*.json"))
 
     query = (raw_stream
         .writeStream
@@ -112,6 +100,6 @@ async def job_task(spark):
 async def test_cancel(spark, azurite):
     task = asyncio.create_task(job_task(spark))
     #await task
-    await asyncio.sleep(10)
+    await asyncio.sleep(30)
     task.cancel()
     print("Done")
